Simplified API calls.
